# src/main/java/com/employee/employee_tax/dto/EmployeeDTO.java
package com.employee.employee_tax.dto;

import jakarta.validation.constraints.*;
import lombok.Getter;
import lombok.Setter;

import java.time.LocalDate;
import java.util.List;
@Setter
@Getter
public class EmployeeDTO {

    private String employeeId;
    private String firstName;
    private String lastName;
    private String email;
    private List<String> phoneNumbers;
    private LocalDate doj;
    private Double salary;

    // Getters and Setters
}