# src/main/resources/api-docs.yaml
openapi: 3.0.3
info:
  title: Employee Tax API
  description: API for managing employees and calculating tax deductions.
  version: 1.0.0
servers:
  - url: http://localhost:8081
    description: Local Server
paths:
  /api/employees:
    post:
      summary: Create a new employee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmployeeDTO'
      responses:
        "201":
          description: Employee created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Employee'
        "400":
          description: Invalid input
        "500":
          description: Server error
  /api/employees/{employeeId}/tax-deductions:
    get:
      summary: Get tax deductions for an employee
      parameters:
        - name: employeeId
          in: path
          required: true
          description: ID of the employee
          schema:
            type: string
      responses:
        "200":
          description: Tax deductions retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaxDeduction'
        "404":
          description: Employee not found
        "500":
          description: Server error
components:
  schemas:
    EmployeeDTO:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        monthlySalary:
          type: number
          format: double
        dateOfJoining:
          type: string
          format: date
    Employee:
      type: object
      properties:
        id:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        monthlySalary:
          type: number
          format: double
        dateOfJoining:
          type: string
          format: date
    TaxDeduction:
      type: object
      properties:
        employeeId:
          type: string
        yearlySalary:
          type: number
          format: double
        taxAmount:
          type: number
          format: double
        cessAmount:
          type: number
          format: double